Assume EDITOR starting with (g)vim is Vim
Rather than assuming the default text editor configured with the
environment variable `VISUAL` or `EDITOR` is Vim if its name is `vim`
or `gvim`, assume it's Vim if its name begins with `vim` or `gvim`.

This is primarily a self-serving change, allowing me to have
`vim-pager` default to opening Vim via my `vim-try-x` wrapper script.